Add thumbnail and downsampled image classes
