+++ web_app/classes/__init__.py
@@ -1,0 +1,1 @@
+from .image_generator import ImageGenerator
