+++ web_app/models/__init__.py
@@ -1,4 +1,6 @@
 from .keyword import Keyword
+from .thumbnail_image import ThumbnailImage
+from .downsampled_image import DownsampledImage
 from .image import Image
 from .album import Album
 from .portfolio import Portfolio
